Move ConditionalUnit into a separate file
