+++ mpd.go
@@ -4,10 +4,8 @@ package mpd
 import (
 	"bytes"
 	"encoding/xml"
-	"fmt"
 	"io"
 	"regexp"
-	"strconv"
 )
 
 // http://mpeg.chiariglione.org/standards/mpeg-dash
@@ -15,49 +13,6 @@ import (
 // http://standards.iso.org/ittf/PubliclyAvailableStandards/MPEG-DASH_schema_files/DASH-MPD.xsd
 
 var emptyElementRE = regexp.MustCompile(`></[A-Za-z]+>`)
-
-// ConditionalUint (ConditionalUintType) defined in XSD as a union of unsignedInt and boolean.
-type ConditionalUint struct {
-	u *uint64
-	b *bool
-}
-
-// MarshalXMLAttr encodes ConditionalUint.
-func (c ConditionalUint) MarshalXMLAttr(name xml.Name) (xml.Attr, error) {
-	if c.u != nil {
-		return xml.Attr{Name: name, Value: strconv.FormatUint(*c.u, 10)}, nil
-	}
-
-	if c.b != nil {
-		return xml.Attr{Name: name, Value: strconv.FormatBool(*c.b)}, nil
-	}
-
-	// both are nil - no attribute, client will threat it like "false"
-	return xml.Attr{}, nil
-}
-
-// UnmarshalXMLAttr decodes ConditionalUint.
-func (c *ConditionalUint) UnmarshalXMLAttr(attr xml.Attr) error {
-	u, err := strconv.ParseUint(attr.Value, 10, 64)
-	if err == nil {
-		c.u = &u
-		return nil
-	}
-
-	b, err := strconv.ParseBool(attr.Value)
-	if err == nil {
-		c.b = &b
-		return nil
-	}
-
-	return fmt.Errorf("ConditionalUint: can't UnmarshalXMLAttr %#v", attr)
-}
-
-// check interfaces
-var (
-	_ xml.MarshalerAttr   = ConditionalUint{}
-	_ xml.UnmarshalerAttr = &ConditionalUint{}
-)
 
 // MPD represents root XML element.
 type MPD struct {
